refactor: remove custom path logic
